+++ examples/create-relayer-key/index.js
@@ -3,7 +3,12 @@ require('dotenv').config();
 const { Defender } = require('@openzeppelin/defender-sdk');
 
 async function main() {
-  const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
+  const creds = {
+    apiKey: process.env.API_KEY,
+    apiSecret: process.env.API_SECRET,
+    //optional https config to keep connection alive. You can pass any configs that are accepted by https.Agent
+    httpsAgent: https.Agent({ keepAlive: true }),
+  };
   const client = new Defender(creds);
 
   const relayerKey = await client.relay.createKey('e65451f5-0f76-4961-8d29-33504319c5f8');