+++ examples/deploy-contract/index.js
@@ -6,7 +6,12 @@ const { AbiCoder } = require('ethers');
 const artifactFile = require('./artifacts/Box.json');
 
 async function main() {
-  const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
+  const creds = {
+    apiKey: process.env.API_KEY,
+    apiSecret: process.env.API_SECRET,
+    //optional https config to keep connection alive. You can pass any configs that are accepted by https.Agent
+    httpsAgent: https.Agent({ keepAlive: true }),
+  };
   const client = new Defender(creds);
 
   // await client.deploy.createBlockExplorerApiKey({