+++ packages/relay-signer/jest.config.js
@@ -1,1 +1,5 @@
-module.exports = require('../../jest.config');
+module.exports = {
+  // https://github.com/jestjs/jest/issues/11617
+  maxWorkers: 1,
+  ...require('../../jest.config'),
+};