+++ examples/relayer-web3-provider/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @openzeppelin/defender-sdk-example-web3-provider
 
+## 1.3.0
+
+### Minor Changes
+
+- 49c1994: refactor: Deprecate optimism-goerli network
+  chore: Enable zkEVM networks
+  chore: Upgrade npm packages
+
+### Patch Changes
+
+- Updated dependencies [49c1994]
+  - @openzeppelin/defender-sdk-relay-signer-client@1.11.0
+
 ## 1.2.0
 
 ### Minor Changes